requirements.txt
numpy==2.0.2
pandas==2.2.3
pydantic==2.10.1
requests==2.32.3
scikit-learn==1.7.1
pyyaml==6.0.2
tqdm==4.67.0
faiss-cpu==1.12.0
sentence_transformers==5.1.0
datasets==4.0.0
ragas==0.3.3
rank_bm25==0.3.3
matplotlib==3.10.6